requirements.txt
absl-py==2.1.0
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
alchemy==20.5
altair==4.2.2
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
assemblyai==0.21.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
python-dotenv
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.11.2
bidict==0.22.1
bitarray==2.8.2
bleach==6.0.0
blinker==1.6.2
blis==0.7.11
bqplot==0.12.37
branca==0.6.0
cachetools==5.3.0
catalogue==2.0.10
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
chroma-hnswlib==0.7.3
chromadb==0.4.15
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
colour==0.1.5
comm==0.1.2
confection==0.1.3
constants==0.6.0
contourpy==1.0.7
cryptography==39.0.2
ctransformers==0.2.27
cycler==0.11.0
cymem==2.0.8
cytoolz==0.12.2
daemon==1.2
dataclasses-json==0.6.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
dnspython==2.3.0
docopt==0.6.2
docutils==0.20.1
elevenlabs==0.2.27
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
entrypoints==0.4
eth-abi==4.2.1
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.0
eth-utils==2.2.2
executing==1.2.0
faiss-cpu==1.8.0
fastapi==0.98.0
fastjsonschema==2.16.2
ffmpy==0.3.0
filelock==3.0.12
Fiona==1.9.1
Flask==2.2.3
flask-babel==3.1.0
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
Flask-SocketIO==5.3.5
flatbuffers==23.5.26
folium==0.13.0
fonttools==4.39.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.5.4
gdown==4.7.1
gemini-api==0.1.6
geographiclib==2.0
geojson==3.0.1
geopandas==0.12.2
geopy==2.3.0
gevent==24.2.1
gitdb==4.0.10
GitPython==3.1.31
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.23.3
google-generativeai==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.61.0
gradio==3.35.2
gradio_client==0.2.7
greenlet==3.0.1
grpcio==1.62.1
grpcio-status==1.62.1
gTTS==2.5.1
h11==0.14.0
h5py==3.10.0
hexbytes==0.3.1
httpcore==0.17.2
httptools==0.6.1
httpx==0.24.1
huggingface-hub==0.21.4
humanfriendly==10.0
idna==2.8
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipyevents==2.0.1
ipyfilechooser==0.6.0
ipykernel==6.21.1
ipyleaflet==0.17.2
ipython==8.12.3
ipython-genutils==0.2.0
ipytree==0.2.2
ipywidgets==7.7.5
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonify==0.5
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter-ydoc==0.2.2
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_fileid==0.6.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.4
jupyterlab_server==2.19.0
kafka-python==2.0.2
keras==3.1.0
kiwisolver==1.4.4
kq==2.2.1
kubernetes==28.1.0
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-experimental==0.0.55
langchain-text-splitters==0.0.1
langcodes==3.3.0
langsmith==0.1.31
leaflet==0.0.3
leafmap==0.18.8
libclang==16.0.6
linkify-it-py==2.0.2
lockfile==0.12.2
lru-dict==1.2.0
Markdown==3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mediapipe==0.10.9
mistune==2.0.5
ml-dtypes==0.3.2
monotonic==1.6
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.4
munch==2.5.0
murmurhash==1.0.10
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.3.0
namex==0.0.7
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.16.2
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.2.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.2.2
onnxruntime==1.16.1
openai==0.27.8
opencv-contrib-python==4.9.0.80
opencv-python==4.8.1.78
openssl-python==0.1.1
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opt-einsum==3.3.0
optree==0.10.0
orjson==3.9.15
overrides==7.4.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.0
parsimonious==0.9.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.5.0
platformdirs==3.0.0
plotly==5.15.0
posthog==3.0.2
preshed==3.0.9
prometheus-client==0.16.0
prompt-toolkit==3.0.36
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.4
pulsar-client==3.3.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.14
pydantic_core==2.16.2
pydeck==0.8.1b0
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.14.0
pymongo==4.4.1
Pympler==1.0.1
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyperclip==1.8.2
PyPika==0.48.9
pyproj==3.4.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline3==3.4.1
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.30
pyshp==2.3.1
PySocks==1.7.1
pystac==1.7.1
pystac-client==0.6.1
python-box==7.0.1
python-daemon==2.2.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-json-logger==2.0.4
python-multipart==0.0.6
python-socketio==5.8.0
pytweening==1.0.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyunormalize==15.0.0
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rlp==3.0.0
rsa==4.9
safetensors==0.4.2
sanic==19.6.0
scikit-learn==1.2.2
scipy==1.10.1
scooby==0.7.1
seaborn==0.12.2
semantic-version==2.10.0
Send2Trash==1.8.0
sentence-transformers==2.5.1
setuptools-scm==8.0.4
shapely==2.0.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.3.2.post1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.1
SQLAlchemy==2.0.22
srsly==2.4.8
stack-data==0.6.2
starlette==0.27.0
streamlit==1.28.0
streamlit-chat==0.1.1
streamlit-lottie==0.0.5
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-hub==0.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.17.1
tf_keras==2.16.0
thinc==8.2.1
threadpoolctl==3.1.0
tiktoken==0.5.1
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.2.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
traittypes==0.2.1
transformers==4.39.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==4.3
uc-micro-py==1.0.2
uri-template==1.2.0
urllib3==1.25.11
uvicorn==0.22.0
validators==0.20.0
wasabi==1.1.2
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.6
weasel==0.3.3
web3==6.11.1
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.2.3
whitebox==2.3.1
whiteboxgui==2.3.0
widgetsnbextension==3.6.4
wrapt==1.15.0
xyzservices==2023.2.0
y-py==0.5.5
yarg==0.1.9
yarl==1.9.2
ypy-websocket==0.8.2
zipp==3.12.1
zope.event==5.0
zope.interface==6.2
